+++ README.md
@@ -1,3 +1,28 @@
 # php-oop-stackoverflow-clone
 php-oop-stackoverflow-clone as oop php project of team learn web programming with shohan
 
+# For the Newly Joined Member
+In this project. We are following this https://docs.php.earth/faq/misc/structure/ folder structure.
+
+
+# Branch
+
+Do your assigned task on the specific branch. Example, You are assinged on Authentication Branch then do your code on Authentication Branch
+
+ # Class Naming
+
+The Class naming Sturcture will be like : YournameClassname example (AtikQuery).
+
+If your task is Authentication releated then You will do it on src\Authentication folder.
+If your task is Database releated then You will do it on src\Database folder.
+
+# Some Essentialts things you need to know
+
+In this project for now, We are not doing any Frontend or UI releated work. We are just doing the backend! After finishing backend, we will impliment UI and Frontend code to Use this App. So It is Requested to do not do any Frontend or UI Related code.
+
+Dont Forget to use the autload and namespace. See others code you will get Idea of the namespace and autoload. If you have any query please ask to our discord general channel.
+
+
+
+
+
