+++ src/Role_Permission/index.php
@@ -1,9 +1,8 @@
 <?php
 
-
-
 namespace Src\Role_Permission;
 
+session_start();
 
 use Src\Authentication\FaiyazRoleBasedAuth;
 
@@ -11,18 +10,20 @@ include_once '../../autoload.php';
 
 
 $user = new FaiyazRoleBasedAuth;
-$login = $user->login('Admin', 'Pass1436');
+$login = $user->login('User', 'Pass1436');
 
-$crud = new FaiyazRolePermission();
-$crud->DeletePost();
 
+// $crud = new FaiyazRolePermission();
+// var_dump ($crud->getPermissionId());
+
+echo "<br>";
 
 
 
  echo "<br>";
 
- if($user->checkRole() == 1){
-     echo "You are Admin";
- }else{
-     echo "You are user";
- }
+//  if($crud->checkRole() == 1){
+//      echo "You are Admin";
+//  }else{
+//      echo "You are user";
+//  }
